requirements.txt
pyscreenshot==2.2
PyAutoGUI==0.9.52
numpy==1.18.1
Pillow==8.0.1